Do not remove parethensis That's potentially risky move on real codebases, but that's not captured anywhere in tests, so I will move ahead with a change.
Would be good if somebody test that on some larger codebase and check how it works for them.

Most other changes is only in testing infra. I take liberty and do following change operating from following premises.
- there only 2 radically different output formats: CSharp and Xml
- Most codegeneration would be platform neutral.
- You will need to override tests based on platform needs (like it is right now)

I do trying not create sets of identical tests for each platform, since it's a bit boring and timeconsuming to validate. For example using existing approach not clear what's platform-independent behaviour and what is OS specific behavior of ClangSharp. After this change all platform independent tests may sit in Base folder, and additional tests and/or augmentations will be in same location as it is now.

Fixes #357